# Refactor translation