# Refactor translation code

A quick optimization pass over the translation code.

## Types of changes

This PR implements the following changes:

* [X] :sparkles: New feature (A non-breaking change which adds functionality).
* [ ] :bug: Bug fix (A non-breaking change, which fixes an issue).
* [ ] :boom: Breaking change (fix or feature that would cause existing functionality to change).

__Changes__
* Optimization of existing code (see benchmarks below)
* Added method `translate!(::LongAminoAcidSeq, ::LongNucleotideSeq); kwargs...`, which avoids allocations. Same kwargs as existing `translate`. Useful for translating small sequences quickly.
* Added method `translate(::Union{DNACodon, RNACodon}, ::GeneticCode)` which efficiently translates a codon. The code defaults to the standard genetic code. This method does no checking, i.e. if you create an invalid genetic code (including nonsense amino acids), this will return those amino acids.

None of the above changes are breaking.

__Benchmarks__
Translating 100 sequences of length `3 * rand(100:3000)`. Mean times:

```
Before:
NucleicAcidAlphabet{4}: 1924 µs
NucleicAcidAlphabet{2}: 2225 µs

After:
NucleicAcidAlphabet{4}: 767 µs  (2.5 x)
NucleicAcidAlphabet{2}: 714 µs  (3.1 x)

After (in-place translation):
NucleicAcidAlphabet{4}: 669 µs (2.9 x)
NucleicAcidAlphabet{2}: 575 µs (3.9 x)
```

## :ballot_box_with_check: Checklist

- [X] :art: The changes implemented is consistent with the [julia style guide](https://docs.julialang.org/en/v1/manual/style-guide/).
- [X] :blue_book: I have updated and added relevant docstrings, in a manner consistent with the [documentation styleguide](https://docs.julialang.org/en/v1/manual/documentation/).
- [X] :blue_book: I have added or updated relevant user and developer manuals/documentation in `docs/src/`.
- [X] :ok: There are unit tests that cover the code changes I have made.
- [X] :ok: The unit tests cover my code changes AND they pass.
- [X] :pencil: I have added an entry to the `[UNRELEASED]` section of the manually curated `CHANGELOG.md` file for this repository.
- [X] :ok: All changes should be compatible with the latest stable version of Julia.
- [X] :thought_balloon: I have commented liberally for any complex pieces of internal code.